requirements.txt
uvicorn
fastapi
odmantic
pydantic
pydantic_settings
email-validator
starlette
motor
passlib[bcrypt]
python-jose[cryptography]
fastapi[all]
pytest
httpx